Welp. Looks like I need to find a way to see if the cursor is within the hitboxes, which would require unprojecting coordinates and all that stuff. Meh, save that for tomorrow xD

+++ core/src/com/games/fnaf/MathStuff.java
@@ -4,8 +4,8 @@ import com.badlogic.gdx.Gdx;
 
 public class MathStuff
 {
-	public static float reverseYCoords(float y)
+	public static int reverseYCoords(float y)
 	{
-		return Math.abs(y - Gdx.graphics.getHeight());
+		return (int)Math.abs(y - Gdx.graphics.getHeight());
 	}
 }
